CPR-305 convert WebClientTest to use multiple instance test base (although it will only hit one application)
